Fix watch mode not reloading changed files
